Cleanup placement of utilities left behind by the core re-org
- NewNodeUUID moves to core next to the NodeUUID type
- ExtractJSON moves to utils as a generic JSON extraction utility
- flows.ContactQueryEscaping becomes contactql.EscapeValue so query
  quoting rules live with the query syntax
- LLMResponse and AirtimeTransfer move out of core/services.go (which
  held no services) into llm.go and a new airtime.go
