+++ core/airtime.go
@@ -5,13 +5,6 @@ import (
 	"github.com/shopspring/decimal"
 )
 
-// LLMResponse is the response from an LLM service call
-type LLMResponse struct {
-	Output       string
-	TokensInput  int64
-	TokensOutput int64
-}
-
 // AirtimeTransfer is the result of an attempted airtime transfer
 type AirtimeTransfer struct {
 	ExternalID string // provider transaction id, when known after initiation
